arch-arm: Define a new misc_info.hh file
This new MiscReg specific couple of source/header file
is supposed to hold informations about MiscRegs. It is
supposed to be more hidden than the misc.hh file, which
contains the definition of MiscRegIndexes and therefore
is the interface to system registers
