+++ src/arch/arm/SConscript
@@ -102,6 +102,7 @@ Source("freebsd/se_workload.cc", tags=['arm isa'])
 Source("fs_workload.cc", tags=['arm isa'])
 Source("regs/int.cc", tags=['arm isa'])
 Source("regs/misc.cc", tags=['arm isa'])
+Source("regs/misc_info.cc", tags=['arm isa'])
 Source("mmu.cc", tags=['arm isa'])
 Source("mpam.cc", tags=['arm isa'])
 Source("nativetrace.cc", tags=['arm isa'])
