+++ src/arch/arm/isa/includes.isa
@@ -56,6 +56,7 @@ output header {{
 #include "arch/arm/insts/mem.hh"
 #include "arch/arm/insts/mem64.hh"
 #include "arch/arm/insts/misc.hh"
+#include "arch/arm/regs/misc_info.hh"
 #include "arch/arm/insts/misc64.hh"
 #include "arch/arm/insts/mult.hh"
 #include "arch/arm/insts/neon64_mem.hh"
@@ -113,6 +114,7 @@ output exec {{
 #include "arch/arm/htm.hh"
 #include "arch/arm/pauth_helpers.hh"
 #include "arch/arm/reg_abi.hh"
+#include "arch/arm/regs/misc_info.hh"
 #include "arch/arm/semihosting.hh"
 #include "arch/arm/utility.hh"
 #include "arch/generic/memhelpers.hh"
